move About Tab into view folder
there is no controller defined for this tab, because it doesn't really need one
