+++ CMakeLists.txt
@@ -23,7 +23,7 @@ set(
   ${PROJECT_SOURCE_DIR}/tabs/view/processes.cc
   ${PROJECT_SOURCE_DIR}/tabs/view/logs.cc
   ${PROJECT_SOURCE_DIR}/tabs/file_chooser.cc
-  ${PROJECT_SOURCE_DIR}/tabs/about.cc
+  ${PROJECT_SOURCE_DIR}/tabs/view/about.cc
 )
 
 ### All the code globbed together for easy formatting ###
