+++ resources/about.glade
@@ -98,23 +98,7 @@ You can monitor processes with profiles under the 'Processes' tab.</property>
           </packing>
         </child>
         <child>
-          <object class="GtkButton" id="link_btn">
-            <property name="label" translatable="yes">For more info click here!</property>
-            <property name="visible">True</property>
-            <property name="can_focus">True</property>
-            <property name="receives_default">True</property>
-            <property name="halign">center</property>
-            <property name="valign">baseline</property>
-            <property name="margin_left">5</property>
-            <property name="margin_right">10</property>
-            <property name="margin_top">5</property>
-            <property name="margin_bottom">5</property>
-          </object>
-          <packing>
-            <property name="expand">False</property>
-            <property name="fill">True</property>
-            <property name="position">1</property>
-          </packing>
+          <placeholder/>
         </child>
       </object>
       <packing>
